requirements.txt
# pipreqs ./src --force and move the requirements.txt to the root of the projects
ImageHash==4.2.1
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.0.1
